Trait diem_parallel_executor::task::ExecutorTask

pub trait ExecutorTask: Sync {
    type T: Transaction;
    type Output: TransactionOutput<T = Self::T>;
    type Error: Clone + Send + Sync;
    type Argument: Sync + Copy;
    fn init(args: Self::Argument) -> Self;

    fn execute_transaction(
        &self, 
        view: &MVHashMapView<'_, <Self::T as Transaction>::Key, <Self::T as Transaction>::Value>, 
        txn: &Self::T
    ) -> ExecutionStatus<Self::Output, Self::Error>;
}

Trait for single threaded transaction executor.

Associated Types

type T: Transaction

Type of transaction and its associated key and value.

type Output: TransactionOutput<T = Self::T>

The output of a transaction. This should contain the side effect of this transaction.

type Error: Clone + Send + Sync

Type of error when the executor failed to process a transaction and needs to abort.

type Argument: Sync + Copy

Type to intialize the single thread transaction executor. Copy and Sync are required because we will create an instance of executor on each individual thread.

Required methods

fn init(args: Self::Argument) -> Self

Create an instance of the transaction executor.

fn execute_transaction(
&self,
view: &MVHashMapView<'_, <Self::T as Transaction>::Key, <Self::T as Transaction>::Value>,
txn: &Self::T
) -> ExecutionStatus<Self::Output, Self::Error>

Execute one single transaction given the view of the current state.