Trait diem_parallel_executor::task::TransactionOutput

pub trait TransactionOutput: Send + Sync {
    type T: Transaction;
    fn get_writes(
        &self
    ) -> Vec<(<Self::T as Transaction>::Key, <Self::T as Transaction>::Value)>;

    fn skip_output() -> Self;
}

Trait for execution result of a transaction.

Associated Types

type T: Transaction

Type of transaction and its associated key and value.

Required methods

fn get_writes(
&self
) -> Vec<(<Self::T as Transaction>::Key, <Self::T as Transaction>::Value)>

Get the side effect of a transaction from its output.

fn skip_output() -> Self

Execution output for transactions that comes after SkipRest signal.